+++ src/PortalBase/StorageKey/PortalNodeKeyCollection.php
@@ -6,12 +6,26 @@ namespace Heptacom\HeptaConnect\Portal\Base\StorageKey;
 
 use Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
+use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\StorageKeyInterface;
 
 /**
  * @extends AbstractCollection<PortalNodeKeyInterface>
  */
 class PortalNodeKeyCollection extends AbstractCollection
 {
+    public function contains($value): bool
+    {
+        return $this->containsByEqualsCheck(
+            $value,
+            static fn (StorageKeyInterface $a, StorageKeyInterface $b): bool => $a->equals($b)
+        );
+    }
+
+    public function unique(): self
+    {
+        return $this->uniqueByContains();
+    }
+
     protected function isValidItem($item): bool
     {
         /* @phpstan-ignore-next-line treatPhpDocTypesAsCertain checks soft check but this is the hard check */