+++ src/StorageBase/FileReferenceRequestKeyCollection.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Storage\Base;
 
 use Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection;
+use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\StorageKeyInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\FileReferenceRequestKeyInterface;
 
 /**
@@ -12,6 +13,19 @@ use Heptacom\HeptaConnect\Storage\Base\Contract\FileReferenceRequestKeyInterface
  */
 final class FileReferenceRequestKeyCollection extends AbstractCollection
 {
+    public function contains($value): bool
+    {
+        return $this->containsByEqualsCheck(
+            $value,
+            static fn (StorageKeyInterface $a, StorageKeyInterface $b): bool => $a->equals($b)
+        );
+    }
+
+    public function unique(): self
+    {
+        return $this->uniqueByContains();
+    }
+
     protected function isValidItem($item): bool
     {
         /* @phpstan-ignore-next-line treatPhpDocTypesAsCertain checks soft check but this is the hard check */